Moved a few functions into debugger.py where they belonged
